Use git binary for pushing tags to avoid authentication issues
This is a temporary fix. Ideally cdx will only fallback to the git binary
if it fails to find the correct credentials or they are not provided.

Currently a hack is used to override the use of the git binary and/or
provide manual basic auth credentials. This should be done instead through
discovery and credential helpers.