Remove basic auth and no git environment variable override.
There is now a requirement to use git when the push (--push) flag is used with
release or promote.

This implementation and requirement significantly simplifies pushing tags for
git as it does not require implementing the configuration required for using
or providing SSH keys or basic authentication. However it likely breaks
Windows compatibility.

As this tool is primarily intended to run on CI/CD systems which are often
containerized or virtualized, this breaking change should not cause any
harm.