Add more tests for partial and overlapping names when scanning tags
This is primarily to make sure multiple similar module names do not
conflate versions. This was partially tested already through the partial
matching tests. The new tests match the more realistic use case of having
2 modules with one containing the same name as the other but with a suffix
or prefix.

This locks in the regular expression behavior for both HEAD and entire
history scanning.