Add preliminary annotated tag support for latest command
Currently this is not fully tested for all cases. Previously annotated
tags would only work for the non-HEAD latest command as there was no
need to check the Git hash.

Annotated tag support now works for the latest command with the --head
flag. This is because annotated tags are now resolved to the object
and the commit reference the tag points to to is used instead of the
annotated tag's reference itself for checking the commit is HEAD.