Use go-cache as backing database for sharing commits quickly
This is ideal if the cdx share server is running as one instance as the
data does not need to be persistent. If the server needs to be
distributed another choice would likely be better.

This should be hidden behind an interface so it can be replaced with
a fake.