Set up initial deployment of a cdx server and allow for target override
This also changes the default for GRPC to be insecure. This is only a
temporary change. As the cdx share feature is undocumented and WIP it is
not expected to be used.

The insecure flag is enabled by default for convenience when
communicating with the default server. When the default server has
correct certificates setup with a real domain, the insecure flag will be
off by default.